Refactor cyphers to own file
